requirements.txt
# Audiofeed Python Microservice
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
pydantic-settings>=2.0

# Podcast (dialogue) — sync, run via run_in_executor
podcastfy>=0.4.0

# HTML stripping (podcastfy dependency; we use it too)
beautifulsoup4>=4.12.0

# LLM + TTS (narration / Instagram)
openai>=1.0.0
google-cloud-texttospeech>=2.14.0
google-generativeai>=0.8.0

# Storage
boto3>=1.35.0

# Webhook
httpx>=0.27.0

# Env
python-dotenv>=1.0.0

# Audio duration (MP3)
mutagen>=1.47.0
